+++ lib/src/_core/network/osu_public_authorization_interceptor.dart
@@ -17,6 +17,7 @@ final class OsuPublicAuthorizationInterceptor
       request.method == RestMethod.get &&
       (request.uri.path.startsWith('/api/v2/users/') ||
           request.uri.path == '/api/v2/spotlights' ||
+          RegExp(r'^/api/v2/news(/[1-9][0-9]*)?$').hasMatch(request.uri.path) ||
           RegExp(
             r'^/api/v2/rankings/(osu|taiko|fruits|mania)/(performance|score|charts)$',
           ).hasMatch(request.uri.path) ||